PiCameraServer/requirements.txt
paho-mqtt
Pillow
python-dotenv